backend-services/ota-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
redis[asyncio]==5.0.4
paho-mqtt==2.0.0
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.84
structlog==24.1.0
prometheus-client==0.20.0

# --- test dependencies ---
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
